+++ src/OpenTelemetry.SemanticConventions/Attributes/AwsAttributes.cs
@@ -15,6 +15,16 @@ namespace OpenTelemetry.SemanticConventions;
 public static class AwsAttributes
 {
     /// <summary>
+    /// The unique identifier of the AWS Bedrock Guardrail. A <a href="https://docs.aws.amazon.com/bedrock/latest/userguide/guardrails.html">guardrail</a> helps safeguard and prevent unwanted behavior from model responses or user messages.
+    /// </summary>
+    public const string AttributeAwsBedrockGuardrailId = "aws.bedrock.guardrail.id";
+
+    /// <summary>
+    /// The unique identifier of the AWS Bedrock Knowledge base. A <a href="https://docs.aws.amazon.com/bedrock/latest/userguide/knowledge-base.html">knowledge base</a> is a bank of information that can be queried by models to generate more relevant responses and augment prompts.
+    /// </summary>
+    public const string AttributeAwsBedrockKnowledgeBaseId = "aws.bedrock.knowledge_base.id";
+
+    /// <summary>
     /// The JSON-serialized value of each item in the <c>AttributeDefinitions</c> request field.
     /// </summary>
     public const string AttributeAwsDynamodbAttributeDefinitions = "aws.dynamodb.attribute_definitions";
@@ -170,12 +180,22 @@ public static class AwsAttributes
     public const string AttributeAwsExtendedRequestId = "aws.extended_request_id";
 
     /// <summary>
+    /// The name of the AWS Kinesis <a href="https://docs.aws.amazon.com/streams/latest/dev/introduction.html">stream</a> the request refers to. Corresponds to the <c>--stream-name</c> parameter of the Kinesis <a href="https://docs.aws.amazon.com/cli/latest/reference/kinesis/describe-stream.html">describe-stream</a> operation.
+    /// </summary>
+    public const string AttributeAwsKinesisStreamName = "aws.kinesis.stream_name";
+
+    /// <summary>
     /// The full invoked ARN as provided on the <c>Context</c> passed to the function (<c>Lambda-Runtime-Invoked-Function-Arn</c> header on the <c>/runtime/invocation/next</c> applicable).
     /// </summary>
     /// <remarks>
     /// This may be different from <c>cloud.resource_id</c> if an alias is involved.
     /// </remarks>
     public const string AttributeAwsLambdaInvokedArn = "aws.lambda.invoked_arn";
+
+    /// <summary>
+    /// The UUID of the <a href="https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-lambda-eventsourcemapping.html">AWS Lambda EvenSource Mapping</a>. An event source is mapped to a lambda function. It's contents are read by Lambda and used to trigger a function. This isn't available in the lambda execution context or the lambda runtime environtment. This is going to be populated by the AWS SDK for each language when that UUID is present. Some of these operations are Create/Delete/Get/List/Update EventSourceMapping.
+    /// </summary>
+    public const string AttributeAwsLambdaResourceMappingId = "aws.lambda.resource_mapping.id";
 
     /// <summary>
     /// The Amazon Resource Name(s) (ARN) of the AWS log group(s).
@@ -295,6 +315,31 @@ public static class AwsAttributes
     /// </ul>
     /// </remarks>
     public const string AttributeAwsS3UploadId = "aws.s3.upload_id";
+
+    /// <summary>
+    /// The ARN of the Secret stored in the Secrets Mangger.
+    /// </summary>
+    public const string AttributeAwsSecretsmanagerSecretArn = "aws.secretsmanager.secret.arn";
+
+    /// <summary>
+    /// The ARN of the AWS SNS Topic. An Amazon SNS <a href="https://docs.aws.amazon.com/sns/latest/dg/sns-create-topic.html">topic</a> is a logical access point that acts as a communication channel.
+    /// </summary>
+    public const string AttributeAwsSnsTopicArn = "aws.sns.topic.arn";
+
+    /// <summary>
+    /// The URL of the AWS SQS Queue. It's a unique identifier for a queue in Amazon Simple Queue Service (SQS) and is used to access the queue and perform actions on it.
+    /// </summary>
+    public const string AttributeAwsSqsQueueUrl = "aws.sqs.queue.url";
+
+    /// <summary>
+    /// The ARN of the AWS Step Functions Activity.
+    /// </summary>
+    public const string AttributeAwsStepFunctionsActivityArn = "aws.step_functions.activity.arn";
+
+    /// <summary>
+    /// The ARN of the AWS Step Functions State Machine.
+    /// </summary>
+    public const string AttributeAwsStepFunctionsStateMachineArn = "aws.step_functions.state_machine.arn";
 
     /// <summary>
     /// The <a href="https://docs.aws.amazon.com/AmazonECS/latest/developerguide/launch_types.html">launch type</a> for an ECS task.