+++ src/OpenTelemetry.SemanticConventions/Attributes/DeploymentAttributes.cs
@@ -17,7 +17,7 @@ public static class DeploymentAttributes
     /// <summary>
     /// 'Deprecated, use <c>deployment.environment.name</c> instead.'.
     /// </summary>
-    [Obsolete("Deprecated, use <c>deployment.environment.name</c> instead.")]
+    [Obsolete("Replaced by <c>deployment.environment.name</c>.")]
     public const string AttributeDeploymentEnvironment = "deployment.environment";
 
     /// <summary>