+++ src/OpenTelemetry.SemanticConventions/Attributes/SystemAttributes.cs
@@ -58,7 +58,7 @@ public static class SystemAttributes
     /// <summary>
     /// Deprecated, use <c>network.connection.state</c> instead.
     /// </summary>
-    [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+    [Obsolete("Replaced by <c>network.connection.state</c>.")]
     public const string AttributeSystemNetworkState = "system.network.state";
 
     /// <summary>
@@ -231,73 +231,73 @@ public static class SystemAttributes
         /// <summary>
         /// close.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string Close = "close";
 
         /// <summary>
         /// close_wait.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string CloseWait = "close_wait";
 
         /// <summary>
         /// closing.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string Closing = "closing";
 
         /// <summary>
         /// delete.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string Delete = "delete";
 
         /// <summary>
         /// established.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string Established = "established";
 
         /// <summary>
         /// fin_wait_1.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string FinWait1 = "fin_wait_1";
 
         /// <summary>
         /// fin_wait_2.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string FinWait2 = "fin_wait_2";
 
         /// <summary>
         /// last_ack.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string LastAck = "last_ack";
 
         /// <summary>
         /// listen.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string Listen = "listen";
 
         /// <summary>
         /// syn_recv.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string SynRecv = "syn_recv";
 
         /// <summary>
         /// syn_sent.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string SynSent = "syn_sent";
 
         /// <summary>
         /// time_wait.
         /// </summary>
-        [Obsolete("Removed, report network connection state with <c>network.connection.state</c> attribute.")]
+        [Obsolete("Replaced by <c>network.connection.state</c>.")]
         public const string TimeWait = "time_wait";
     }